# packages/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Hookwing Webhook API
  version: "0.0.1"
  description: |
    Hookwing is a webhook delivery platform that reliably receives, routes, and delivers webhooks
    with automatic retries, signature verification, and tier-based feature gating.
  contact:
    name: Hookwing Engineering
    url: https://hookwing.com
    email: support@hookwing.com
servers:
  - url: https://api.hookwing.com
    description: Production
  - url: https://api-dev.hookwing.com
    description: Development
security:
  - BearerAuth: []
...
paths:
...
  /v1/endpoints:
    post:
      operationId: createEndpoint
      tags: [Endpoints]
      summary: Create webhook endpoint
      description: |
        Creates a new webhook endpoint with an auto-generated signing secret (shown only once).
        Subject to tier max_destinations limit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndpointCreateInput"
            example:
              url: https://example.com/webhooks
              description: Production endpoint
              eventTypes: ["order.created", "order.updated"]
      responses:
        "201":
          description: Endpoint created (signing secret shown only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndpointWithSecret"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Tier endpoint limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TierLimitError"

    get:
      operationId: listEndpoints
      tags: [Endpoints]
      summary: List all endpoints for workspace
      responses:
        "200":
          description: Endpoints list
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoints:
                    type: array
                    items:
                      $ref: "#/components/schemas/Endpoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/endpoints/{id}:
    get:
      operationId: getEndpoint
      tags: [Endpoints]
      summary: Get endpoint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Endpoint found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateEndpoint
      tags: [Endpoints]
      summary: Update endpoint
      description: Partial update — only provided fields are changed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndpointUpdateInput"
      responses:
        "200":
          description: Endpoint updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteEndpoint
      tags: [Endpoints]
      summary: Delete endpoint
      description: Hard-deletes the endpoint.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Endpoint deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==========================================================================
  # Events
  # ==========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key from signup or key creation. Format - `hk_live_<random>`
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized
            message: Missing Authorization header
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Not found
...
    ValidationError:
      description: Invalid request body
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid input
              details:
                type: object
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...
    TierLimits:
      type: object
      properties:
        max_destinations:
          type: integer
        max_events_per_month:
          type: integer
        max_payload_size_bytes:
          type: integer
        max_retry_attempts:
          type: integer
        retention_days:
          type: integer
        rate_limit_per_second:
          type: integer
    TierFeatures:
      type: object
      properties:
        custom_headers:
          type: boolean
        ip_whitelist:
          type: boolean
        transformations:
          type: boolean
        dead_letter_queue:
          type: boolean
        priority_delivery:
          type: boolean
        webhook_signing:
          type: boolean
        analytics:
          type: boolean
        team_members:
          type: integer
    TierConfig:
      type: object
      properties:
        slug:
          type: string
          example: paper-plane
        name:
          type: string
          example: Paper Plane
        price_monthly_usd:
          type: number
          example: 0
        limits:
          $ref: "#/components/schemas/TierLimits"
        features:
          $ref: "#/components/schemas/TierFeatures"

    # -- Auth --
...
    EndpointCreateInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Must use HTTPS
          example: https://example.com/webhooks
        description:
          type: string
          maxLength: 500
        eventTypes:
          type: array
          items:
            type: string
          description: Event types to subscribe to (null = all)
        fanoutEnabled:
          type: boolean
          default: true
          description: Whether this endpoint receives fan-out events from other endpoints
        metadata:
          type: object
          additionalProperties:
            type: string
    EndpointUpdateInput:
      type: object
      properties:
        url:
          type: string
          format: uri
        description:
          type: string
          maxLength: 500
          nullable: true
        eventTypes:
          type: array
          items:
            type: string
          nullable: true
        isActive:
          type: boolean
        fanoutEnabled:
          type: boolean
          description: Whether this endpoint receives fan-out events from other endpoints
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
    Endpoint:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        url:
          type: string
        description:
          type: string
          nullable: true
        eventTypes:
          type: array
          items:
            type: string
          nullable: true
        fanoutEnabled:
          type: boolean
          description: Whether this endpoint receives fan-out events from other endpoints
        isActive:
          type: boolean
        rateLimitPerSecond:
          type: integer
          nullable: true
        metadata:
          type: object
          nullable: true
        createdAt:
          type: integer
        updatedAt:
          type: integer
    EndpointWithSecret:
      allOf:
        - $ref: "#/components/schemas/Endpoint"
        - type: object
          properties:
            secret:
              type: string
              description: Signing secret (whsec_*). Shown only on creation.
              example: whsec_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6
    TierLimitError:
      type: object
      properties:
        error:
          type: string
          example: Endpoint limit reached
        limit:
          type: integer
        current:
          type: integer
        tier:
          type: string
        upgradePath:
          type: array
          items:
            $ref: "#/components/schemas/TierConfig"

    # -- Events --
...